Functional ish cart modal
CartModal is included directly in the Cart component, and the content has been set to an OrderSummary component. Page jumping has been fixed, as well as weird navigation buttons! Just need to adjust the size and position of the cart overlay and we should be good.
