--- frontend/src/components/CartOverlay.js
@@ -1,16 +1,0 @@
-import '../styles/modal.css'
-
-const CartOverlay = ({ handleClose, show, children }) => {
-    const showHideClassName = show ? "modal display-block" : "modal display-none";
-    console.log("Modal class applied:", showHideClassName);
-
-    return (
-        <div className={showHideClassName} onClick={handleClose}>
-            <section className="modal-main" onClick={(e) => e.stopPropagation()}>
-                {children}
-            </section>
-        </div>
-    );
-};
-
-export default CartOverlay;
