+++ frontend/src/styles/modal.css
@@ -1,10 +1,12 @@
+
+
 .modal {
     position: fixed;
     top: 0;
     left: 0;
     width:100%;
     height: 100%;
-    background: rgba(0, 0, 0, 0.6);
+
 }
 
 .modal-main {
