+++ frontend/src/styles/navbutton.css
@@ -10,7 +10,7 @@
     transition: color 0.3s, text-shadow 0.3s;
 }
 
-.nav-button:hover, :focus, :active {
+.nav-button:hover {
     transform: translateY(-2px);
     color: white;
     text-shadow: 1px 0 var(--main-color), -1px 0 var(--main-color), 0 4px var(--main-color), 0 -1px var(--main-color);
